COMP: Allow build to configure with explicit instantiation off
The sources of many modules only contain explicit instantiation code, these
modules are correctly exported as shared libraries when explicit instantiation
is on but should not be exported when explict instantiation is turned off.
This patch adds a check to the ENABLE_SHARED property in these modules that
only adds it if explicit instantiation is turned on.

Change-Id: Ib38a7310d41d4f5c5b820473af15d9550491f8b3