COMP: Avoid compiler warnings with IPO
When compiling with agressive IPO optimizations,
some invalid code constructs were identified where
the loop iterations did not match the size of the
objects being indexed.

The recent change to have Python compile for 4D cases
identified issues for code bases that are only valid
at exactly 3D code.

The HexahedronCell algorithms are only valid for 3D cases,
but the Spatial Object reader for Mesh instantiates all the
mesh objects (both those that are dimension agnostic, and those
with limited dimension support) to allow a consistent interface
implementation.

The solution here to avoid compiler warnings is to hard-code the
HexahedronCell algorithms to exactly 3D, add Exceptions if the
algorithms are used for other dimensions.