STYLE: Allow use of clang-tidy as part of cmake compilation
Recent versions of cmake allow running clang-tidy as
part of the compilation process.

A <LANG>_CLANG_TIDY target property and supporting
CMAKE_<LANG>_CLANG_TIDY variable were introduced to tell the Makefile
Generators and the Ninja generator to run clang-tidy along with the
compiler for C and CXX languages.  Supported in cmake 3.6+.