WIP: ENH: Added type hints for wrapped Python classes and methods.
Logic was added to the igenerator.py file to produce type hints for wrapped ITK classes in Python.

Over half of the available ITK classes are currently supported by this feature, with more being worked on.

Two files are created, init.pyi and methods.pyi which contain the hint information needed for an IDE like Pycharm. Once built there should be no extra setup needed to get the hinting functionality.

The hints use have Template and Proxy names appended to them to differentiate them from the actual ITK classes. The hints are not perfect, they are merely suggestions for potentially correct input options. For example methods may only accept a double or character depending on the object type, but the hints will suggest both leaving it up to the user to determine which type is valid.