BUG: Restructured python type hint generation procedure
A few inter-submodule dependencies were causing a few classes typehints to fail.
This commit restructures the generation process to allow for these dependencies.
The generation occurs in two stages:
1. pickle files are saved for each class defined for each submodule in igenerator
	(Some classes are defined multiple times (once for each submodule))
2. Pickle files are merged and stub files are produced in pyi_generator
	(This runs once after all runs of igenerator have finished)
Currently pyi_generator runs before igenerator, this is a problem.
A dependency needs to be added to cause pyi_generator to be ran last.

Co-authored-by: Kian Weimer <kianweimer@gmail.com>