COMP: Remove options that are always overridden
The values of ${module_prefix}_WRAP_CASTXML and ${module_prefix}_WRAP_SWIGINTERFACE
are always forced to be the same as ITK_WRAPPING.  ITK_WRAPPING is currently an alias
of ITK_WRAP_PYTHON since python is the only wrapping mechanism supported.

These changes are a first step in greatly simplifying the macro inclusion
logic.