BUG: A temporary work-around for HDF5 testing
The HDF5ImageIO objects are single use for writing.  After
the initial use, the WriteImageInformation function is
short circuited, and can not be reset for writing a second image.

This bug was exposed while fixing a problem with the not
throwing an exception when attempting to write to a path
that does not exist.

This temporary work around in the test framework allows fixing
the initial problem without requiring fixing the HDF5ImageIO
re-use issues.

NOTE: Resetting the HDF5ImageIO object at the end of writing
fixes the re-use problem for non-streaming case, but that
causes the streaming tests to fail.